switch to `tape` for testing
tape works in the browser
